+++ pom.xml
@@ -12,5 +12,13 @@
 	    <version>20180813</version>
 	</dependency>
 
+	<!-- https://mvnrepository.com/artifact/junit/junit -->
+	<dependency>
+	    <groupId>junit</groupId>
+	    <artifactId>junit</artifactId>
+	    <version>4.12</version>
+	    <scope>test</scope>
+	</dependency>
+	
   </dependencies>
 </project>
