--- src/main/java/Messager.java
@@ -1,8 +1,0 @@
-
-public class Messager 
-{
-	public void PrintMessage(String message)
-	{
-		System.out.println(message);
-	}
-}
