+++ src/main/java/com/qa/accounts/Account.java
@@ -1,3 +1,4 @@
+package com.qa.accounts;
 
 public class Account
 {
