+++ src/main/java/com/qa/accounts/Messager.java
@@ -1,0 +1,9 @@
+package com.qa.accounts;
+
+public class Messager 
+{
+	public void PrintMessage(String message)
+	{
+		System.out.println(message);
+	}
+}
